+++ src/components/listHoc.jsx
@@ -12,7 +12,7 @@ import {
   arrayMove,
 } from 'react-sortable-hoc';
 import FoldIcon from './foldIcon';
-import DescriptionList, { getDescription } from './descriptionList1';
+import DescriptionList, { getDescription } from './descriptionList';
 
 const DragHandle = SortableHandle(() => (
   <span className="fr-move-icon">:::</span>
